Adaptive RK integration implemented
An adaptive RK integration scheme is implemented. In this case, the vector b of the tableau is double the length of the explict integration. The first half of the vector are the b values for a high order integration estimate. The second half for a low order error estimate. The difference between the high order estimate and the low order estimate gives the error estimate of the high order result. If the error is greater than a tolerance, the time step is divided by 2*numIterations (so gets smaller faster as more iterations occur).