Added row reduction function
Given an NxM matrix, row reduction first performs the operations required to give a matrix in echelon form. From there, further operations are performed to give reduced row echelon form. This allows for e.g. inverting the matrix, or using matrix operations to solve a series of linear equations.
Note: currently, no checking is performed. It is assumed that the input matrix is able to safely be operated to reduced row echelon form, without any divide by zero errors or whatever. This may be improved at a later date.